fixed until modifica docente

+++ src/main/java/com/example/demo/repository/DocenteRepository.java
@@ -5,4 +5,4 @@ package com.example.demo.repository;
 import com.example.demo.entity.Docente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocenteRepository extends JpaRepository<Docente, Long> {}
+public interface DocenteRepository extends JpaRepository<Docente, Integer> {}
